+++ AlchemistCode/Cards/Common/Backfire.cs
@@ -1,12 +1,8 @@
-using BaseLib.Extensions;
 using BaseLib.Utils;
 using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
-using MegaCrit.Sts2.Core.HoverTips;
-using MegaCrit.Sts2.Core.Localization.DynamicVars;
 using MegaCrit.Sts2.Core.Models.Powers;
-using MegaCrit.Sts2.Core.ValueProps;
 
 namespace Alchemist.AlchemistCode.Cards.Common;
 
@@ -15,17 +11,17 @@ public class Backfire : AlchemistCard
 {
     public Backfire() : base(2, CardType.Attack, CardRarity.Common, TargetType.AnyEnemy)
     {
-        WithDamage(14, 4);
-        WithVar("poison", 2, 0);
+        WithDamage(12, 4);
+        WithVar("SelfPoison", 1, 0);
         WithTip(typeof(PoisonPower));
     }
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
         await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_fire_burst"),
-            sfx: "event:/sfx/characters/attack_fire")
-            .WithAttackerAnim(HeavyAttackAnim, HeavyAttackDelay).Execute(choiceContext);
+            sfx: "event:/sfx/characters/attack_fire").WithAttackerAnim(HeavyAttackAnim, HeavyAttackDelay)
+            .Execute(choiceContext);
         await PowerCmd.Apply<PoisonPower>(choiceContext, Owner.Creature,
-            DynamicVars["poison"].IntValue, Owner.Creature, this);
+            DynamicVars["SelfPoison"].IntValue, Owner.Creature, this);
     }
 }