+++ AlchemistCode/Cards/Common/CausticStrike.cs
@@ -1,26 +1,39 @@
+using Alchemist.AlchemistCode.Powers;
 using BaseLib.Utils;
 using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
-using Alchemist.AlchemistCode.Powers;
+using MegaCrit.Sts2.Core.Localization;
+using MegaCrit.Sts2.Core.Models.Powers;
 
 namespace Alchemist.AlchemistCode.Cards.Common;
 
-[CardTheme(CardTheme.Antitoxin)]
+[CardTheme(CardTheme.Ferment, CardTheme.Poison)]
 public class CausticStrike : AlchemistCard
 {
+    protected override bool Ferments => true;
+
     public CausticStrike() : base(1, CardType.Attack, CardRarity.Common, TargetType.AnyEnemy)
     {
-        WithTags(CardTag.Strike);
-        WithDamage(6, 3);
-        WithVar("antitoxin", 3, 2);
-        WithTip(typeof(AntitoxinPower));
+        WithDamage(6, 2);
+        WithVar("Poison", 1, 1);
+        WithKeyword(CardKeyword.Retain);
+        WithTip(typeof(PoisonPower));
+    }
+
+    private int Ripened => DynamicVars["Poison"].IntValue + FermentTurns;
+
+    protected override void AddExtraArgsToDescription(LocString description)
+    {
+        base.AddExtraArgsToDescription(description);
+        description.Add("Ripened", FermentTurns > 0 ? $" ([green]{Ripened}[/green])" : "");
     }
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
         await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_slime_impact")).Execute(choiceContext);
-        await PowerCmd.Apply<AntitoxinPower>(choiceContext, Owner.Creature,
-            DynamicVars["antitoxin"].IntValue, Owner.Creature, this);
+        if (play.Target is not { IsAlive: true } target) return;
+        PoisonSplash(target);
+        await PowerCmd.Apply<PoisonPower>(choiceContext, target, Ripened, Owner.Creature, this);
     }
 }