+++ AlchemistCode/Cards/Uncommon/Drench.cs
@@ -1,0 +1,35 @@
+using Alchemist.AlchemistCode.Powers;
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.Commands;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using MegaCrit.Sts2.Core.Models.Powers;
+
+namespace Alchemist.AlchemistCode.Cards.Uncommon;
+
+[CardTheme(CardTheme.Poison)]
+public class Drench : AlchemistCard
+{
+    public Drench() : base(2, CardType.Attack, CardRarity.Uncommon, TargetType.AnyEnemy)
+    {
+        WithDamage(13, 4);
+        WithVar("Poison", 2, 0);
+        WithVar("SelfPoison", 1, 0);
+        WithTip(typeof(PoisonPower));
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_heavy_blunt"),
+            tmpSfx: "heavy_attack.mp3").WithAttackerAnim(HeavyAttackAnim, HeavyAttackDelay)
+            .Execute(choiceContext);
+        if (play.Target is { IsAlive: true } target)
+        {
+            PoisonSplash(target);
+            await PowerCmd.Apply<PoisonPower>(choiceContext, target,
+                DynamicVars["Poison"].IntValue, Owner.Creature, this);
+        }
+        await PowerCmd.Apply<PoisonPower>(choiceContext, Owner.Creature,
+            DynamicVars["SelfPoison"].IntValue, Owner.Creature, this);
+    }
+}