+++ AlchemistCode/Cards/Uncommon/Numb.cs
@@ -1,3 +1,4 @@
+using Alchemist.AlchemistCode.Powers;
 using BaseLib.Utils;
 using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
@@ -13,15 +14,14 @@ public class Numb : AlchemistCard
 
     public Numb() : base(2, CardType.Skill, CardRarity.Uncommon, TargetType.Self)
     {
-        WithBlock(14, 4);
-        WithVar("SelfPoison", 3, 0);
+        WithBlock(10, 4);
+        WithPower<NumbPower>(1, 0);
         WithTip(typeof(PoisonPower));
     }
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
         await CommonActions.CardBlock(this, play);
-        await PowerCmd.Apply<PoisonPower>(choiceContext, Owner.Creature,
-            DynamicVars["SelfPoison"].IntValue, Owner.Creature, this);
+        await CommonActions.ApplySelf<NumbPower>(choiceContext, this);
     }
 }